requirements.txt
grpcio
fabric-protos-python
